+++ src/main/java/com/createLines/task/repository/LineSegmentRepository.java
@@ -3,5 +3,5 @@ package com.createLines.task.repository;
 import com.createLines.task.model.LineSegment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LineSegmentRepository extends JpaRepository<LineSegment, Integer> {
+public interface LineSegmentRepository extends JpaRepository<LineSegment, Long> {
 }
